Replace SignUpWithPasswordPayload with UserLogInPayload
